refactor: rewrite logic
BREAKING CHANGE: complete project restructuring
