+++ internal/constants/constants.go
@@ -1,0 +1,4 @@
+package constants
+
+const ComposeLabel = "com.docker.compose.project"
+const FormatedStrPadded = "<code> %-8v</code><code>%v</code>"
