+++ internal/telegram/telegram.go
@@ -1,18 +1,26 @@
-package main
+package telegram
 
 import (
 	"fmt"
+	"math"
 	"time"
 
 	tb "gopkg.in/tucnak/telebot.v2"
 
-	"github.com/rs/zerolog/log"
+	"github.com/docker/docker/api/types"
+	"github.com/mrmarble/teledock/internal/docker"
+	"github.com/rs/zerolog"
+	zero "github.com/rs/zerolog/log"
 )
+
+var log zerolog.Logger
 
 // Telegram represents the telegram bot.
 type Telegram struct {
 	bot                *tb.Bot
+	dckr               docker.Docker
 	handlersRegistered bool
+	admins             []int64
 }
 
 // Command represent a telegram command.
@@ -24,12 +32,14 @@ type Command struct {
 }
 
 // NewBot returns a Telegram bot.
-func NewBot(token string) (*Telegram, error) {
+func NewBot(token string, dckr docker.Docker, admins []int64) (*Telegram, error) {
+	log = zero.With().Str("package", "Telegram").Logger()
+
 	bot, err := tb.NewBot(tb.Settings{
 		Token:  token,
 		Poller: &tb.LongPoller{Timeout: 10 * time.Second},
 		Reporter: func(err error) {
-			log.Error().Str("module", "telegram").Err(err).Msg("telebot internal error")
+			log.Error().Err(err).Msg("telebot internal error")
 		},
 	})
 
@@ -37,16 +47,16 @@ func NewBot(token string) (*Telegram, error) {
 		return nil, err
 	}
 
-	log.Info().Str("module", "telegram").Int64("id", bot.Me.ID).Str("name", bot.Me.FirstName).Str("username", bot.Me.Username).Msg("connected to telegram")
+	log.Info().Int64("id", bot.Me.ID).Str("name", bot.Me.FirstName).Str("username", bot.Me.Username).Msg("connected to telegram")
 
-	return &Telegram{bot: bot}, nil
+	return &Telegram{bot: bot, dckr: dckr, admins: admins}, nil
 }
 
 // Start starts polling for telegram updates.
 func (t *Telegram) Start() {
 	t.registerHandlers()
 
-	log.Info().Str("module", "telegram").Msg("start polling")
+	log.Info().Msg("start polling")
 	t.bot.Start()
 }
 
@@ -56,7 +66,7 @@ func (t *Telegram) registerHandlers() {
 		return
 	}
 
-	log.Info().Str("module", "telegram").Msg("registering handlers")
+	log.Info().Msg("registering handlers")
 
 	// Temporal list to send the commands to telegram
 	botCommandList := []tb.Command{}
@@ -126,14 +136,14 @@ func (t *Telegram) registerHandlers() {
 	t.bot.Handle(tb.OnCallback, t.handleCallback)
 
 	if err := t.bot.SetCommands(botCommandList); err != nil {
-		log.Fatal().Str("module", "telegram").Err(err).Msg("error registering commands")
+		log.Fatal().Err(err).Msg("error registering commands")
 	}
 
 	t.handlersRegistered = true
 }
 
 func (t *Telegram) isSuperAdmin(user *tb.User) bool {
-	for _, uid := range superAdmins {
+	for _, uid := range t.admins {
 		if user.ID == uid {
 			return true
 		}
@@ -164,12 +174,12 @@ func (t *Telegram) send(to tb.Recipient, what interface{}, options ...interface{
 		}
 
 		if try > 5 {
-			log.Error().Str("module", "telegram").Err(err).Msg("send aborted, retry limit exceeded")
+			log.Error().Err(err).Msg("send aborted, retry limit exceeded")
 			return nil
 		}
 
 		backoff := time.Second * 5 * time.Duration(try)
-		log.Warn().Str("module", "telegram").Err(err).Str("sleep", backoff.String()).Msg("send failed, sleeping and retrying")
+		log.Warn().Err(err).Str("sleep", backoff.String()).Msg("send failed, sleeping and retrying")
 		time.Sleep(backoff)
 		try++
 	}
@@ -197,13 +207,80 @@ func (t *Telegram) reply(to *tb.Message, what interface{}, options ...interface{
 		}
 
 		if try > 5 {
-			log.Error().Str("module", "telegram").Err(err).Msg("reply aborted, retry limit exceeded")
+			log.Error().Err(err).Msg("reply aborted, retry limit exceeded")
 			return nil
 		}
 
 		backoff := time.Second * 5 * time.Duration(try)
-		log.Warn().Str("module", "telegram").Err(err).Str("sleep", backoff.String()).Msg("reply failed, sleeping and retrying")
+		log.Warn().Err(err).Str("sleep", backoff.String()).Msg("reply failed, sleeping and retrying")
 		time.Sleep(backoff)
 		try++
+	}
+}
+
+func (t *Telegram) makeContainerMenu(options types.ContainerListOptions, callback string) *tb.ReplyMarkup {
+	buttonsPerRow := 3
+	containers := t.dckr.List(options)
+
+	menu := t.bot.NewMarkup()
+	rowNumber := int(math.Ceil(float64(len(containers)) / float64(buttonsPerRow)))
+	buttons := []tb.InlineButton{}
+	rows := make([][]tb.InlineButton, rowNumber)
+	for index, container := range containers {
+		if index != 0 && index%buttonsPerRow == 0 {
+			rows = append(rows, buttons)
+			buttons = nil
+		}
+
+		btn := menu.Data(container.Names[0][1:], fmt.Sprintf("%v:%v", index, container.ID[:10]), fmt.Sprintf("%v:%v", callback, container.ID[:10])).Inline()
+		buttons = append(buttons, *btn)
+	}
+	if len(buttons) > 0 {
+		rows = append(rows, buttons)
+	}
+	menu.InlineKeyboard = rows
+	return menu
+}
+
+func (t *Telegram) askForContainer(m *tb.Message, listOps types.ContainerListOptions, cb string) {
+	t.reply(m, "Choose a container", t.makeContainerMenu(listOps, cb))
+}
+
+func (t *Telegram) handleLog(c *tb.Callback, payload string) {
+	logs, err := t.dckr.Logs(payload, "10")
+	if err != nil {
+		t.callbackResponse(c, err, payload, "")
+		return
+	}
+	for index, chunk := range logs {
+		if index == 0 {
+			t.callbackResponse(c, err, payload, fmt.Sprintf(FormatedStr, chunk))
+		}
+		if index != 0 && chunk != "" {
+			t.send(c.Message.Chat, fmt.Sprintf(FormatedStr, chunk), tb.ModeHTML)
+		}
+		time.Sleep(100 * time.Millisecond)
+	}
+}
+
+func (t *Telegram) callbackResponse(c *tb.Callback, err error, payload interface{}, response string) {
+	if err != nil {
+		err = t.bot.Respond(c, &tb.CallbackResponse{Text: err.Error(), ShowAlert: false})
+		if err != nil {
+			log.Fatal().Str("module", "utils").Err(err).Msg("error replying to message")
+		}
+		_, err = t.bot.Edit(c.Message, fmt.Sprintf("Container %v errored: %v", payload, err.Error()))
+		if err != nil {
+			log.Fatal().Str("module", "utils").Err(err).Msg("error editing message")
+		}
+	} else {
+		err := t.bot.Respond(c, &tb.CallbackResponse{Text: "", ShowAlert: false})
+		if err != nil {
+			log.Fatal().Str("module", "utils").Err(err).Msg("error replying to message")
+		}
+		_, err = t.bot.Edit(c.Message, response, tb.ModeHTML)
+		if err != nil {
+			log.Fatal().Str("module", "utils").Err(err).Msg("error editing message")
+		}
 	}
 }
